+++ R/mid_terms.R
@@ -48,10 +48,8 @@ mid.terms <- function(
     main.effects <- dots$me
   if (missing(interactions) && !is.null(dots$ie))
     interactions <- dots$ie
-  terms <- ifnot.null(
-    attr(object$terms %||% attr(object, "terms"), "term.labels"),
-    attr(object, "term.labels")
-  )
+  terms <- attr(object$terms %||% attr(object, "terms"), "term.labels")
+  terms <- terms %||% attr(object, "term.labels")
   if (!main.effects)
     terms <- terms[grepl(":", terms)]
   if (!interactions)